+++ test/v2/models/responses/MovementResponseSpec.scala
@@ -22,67 +22,25 @@ import org.scalatest.matchers.must.Matchers
 import org.scalatestplus.scalacheck.ScalaCheckDrivenPropertyChecks
 import play.api.libs.json.JsSuccess
 import play.api.libs.json.Json
-import v2.models.EORINumber
+import v2.models.MessageId
 import v2.models.MovementId
-import v2.models.MovementReferenceNumber
-
-import java.time.OffsetDateTime
-import java.time.ZoneOffset
 
 class MovementResponseSpec extends AnyFreeSpec with Matchers with ScalaCheckDrivenPropertyChecks {
 
-  private val hexId    = Gen.listOfN(16, Gen.hexChar).map(_.mkString.toLowerCase)
-  private val dateTime = OffsetDateTime.of(2022, 8, 15, 11, 30, 0, 0, ZoneOffset.UTC)
+  private val gen = Gen.listOfN(10, Gen.alphaChar).map(_.mkString)
 
-  "when DepartureResponse is serialized, return an appropriate JsObject" in {
-    val departureId = MovementId(hexId.sample.get)
-
-    val actual = MovementResponse.movementResponseFormat.writes(
-      MovementResponse(
-        _id = departureId,
-        enrollmentEORINumber = EORINumber("GB123"),
-        movementEORINumber = EORINumber("GB456"),
-        movementReferenceNumber = Some(MovementReferenceNumber("MRN001")),
-        created = dateTime,
-        updated = dateTime
-      )
-    )
-
-    val expected = Json.obj(
-      "_id"                     -> departureId,
-      "enrollmentEORINumber"    -> "GB123",
-      "movementEORINumber"      -> "GB456",
-      "movementReferenceNumber" -> "MRN001",
-      "created"                 -> "2022-08-15T11:30:00Z", // due to OffsetDateTime in UTC
-      "updated"                 -> "2022-08-15T11:30:00Z"
-    )
-    actual mustBe expected
+  "when MessageID is serialized, return an appropriate JsObject" in forAll(gen, gen) {
+    (movement, message) =>
+      val actual   = MovementResponse.movementResponseFormat.writes(MovementResponse(MovementId(movement), MessageId(message)))
+      val expected = Json.obj("movementId" -> movement, "messageId" -> message)
+      actual mustBe expected
   }
 
-  "when an appropriate JsObject is deserialized, return a DepartureResponse" in {
-    val departureId = MovementId(hexId.sample.get)
-
-    val expected =
-      MovementResponse(
-        _id = departureId,
-        enrollmentEORINumber = EORINumber("GB123"),
-        movementEORINumber = EORINumber("GB456"),
-        movementReferenceNumber = Some(MovementReferenceNumber("MRN001")),
-        created = dateTime,
-        updated = dateTime
-      )
-
-    val actual = MovementResponse.movementResponseFormat.reads(
-      Json.obj(
-        "_id"                     -> departureId,
-        "enrollmentEORINumber"    -> "GB123",
-        "movementEORINumber"      -> "GB456",
-        "movementReferenceNumber" -> "MRN001",
-        "created"                 -> "2022-08-15T11:30:00Z", // due to OffsetDateTime in UTC
-        "updated"                 -> "2022-08-15T11:30:00Z"
-      )
-    )
-    actual mustBe JsSuccess(expected)
+  "when an appropriate JsObject is deserialized, return a MessageId" in forAll(gen, gen) {
+    (movement, message) =>
+      val actual   = MovementResponse.movementResponseFormat.reads(Json.obj("movementId" -> movement, "messageId" -> message))
+      val expected = MovementResponse(MovementId(movement), MessageId(message))
+      actual mustBe JsSuccess(expected)
   }
 
 }
